Changed Makefile

+++ src/ports.c
@@ -12,7 +12,7 @@ void outb(unsigned short port, unsigned char data)
   asm volatile("outb %0, %1" : "=a"(data) : "d"(port));
 }
 
-unsigned long inl(unsigned long port)
+unsigned long inl(unsigned short port)
 {
   unsigned long ret;
   asm volatile("inl %1, %0" : "=a"(ret) : "d"(port));
